backend/pyproject.toml
[project]
name = "companion-backend"
version = "0.1.0"
description = "FastAPI backend for the companion project"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.111.0",
  "uvicorn[standard]>=0.30.0",
  "python-dotenv>=1.0.1",
  "google-generativeai>=0.8.3"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "httpx>=0.27.0"
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
